Add files via upload
This folder contains the before and after files for an identified bug. The first two files are the style sheet and code for the page to which I attempted to add the 3rd party API map with markers. The second two files are the finalised versions of those files with the bug corrected.